Separated reading from Repository Logic
